# api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package api

// Defines values for ErrorCode.
const (
	ERROR ErrorCode = "ERROR"
)

// Error A generic error
type Error struct {
	Code    ErrorCode `json:"code"`
	Message string    `json:"message"`
	Status  int64     `json:"status"`
}

// ErrorCode defines model for ErrorCode.
type ErrorCode string

// SlackEvent A slack event
type SlackEvent struct {
	// Channel The channel id the message was sent from
	Channel string `json:"channel"`

	// EventTs The event timestamp
	EventTs string `json:"event_ts"`

	// Text The content text
	Text string `json:"text"`

	// Ts The message timestamp
	Ts string `json:"ts"`

	// User The user id of the sender
	User string `json:"user"`
}

// BadRequest A generic error
type BadRequest = Error

// ImagineResponse defines model for ImagineResponse.
type ImagineResponse struct {
	// Accepted Determines whether the prompt was accepted
	Accepted bool `json:"accepted"`

	// Challenge An optional field used by Slack for endpoint verification
	Challenge *string `json:"challenge,omitempty"`
}

// InternalServerError A generic error
type InternalServerError = Error

// ImagineJSONBody defines parameters for Imagine.
type ImagineJSONBody struct {
	// Challenge An optional field used by Slack for endpoint verification
	Challenge *string `json:"challenge,omitempty"`

	// Event A slack event
	Event SlackEvent `json:"event"`

	// EventId The event id
	EventId string `json:"event_id"`

	// EventTime The event time
	EventTime int64 `json:"event_time"`

	// TeamId The team id
	TeamId string `json:"team_id"`

	// Type The event type
	Type interface{} `json:"type"`
}

// ImagineJSONRequestBody defines body for Imagine for application/json ContentType.
type ImagineJSONRequestBody ImagineJSONBody