# openapi.yaml
openapi: 3.0.0

info:
  version: 1.0.0
  title: FetchmojiAI Service
  description: A Slack bot service that generates emojis

servers:
  - url: http://localhost:8080/v1

paths:
  /imagine:
    post:
      description: Imagines an emoji
      operationId: imagine
      tags:
        - Imagine
      requestBody:
        description: The request body to imagine an emoji
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                challenge:
                  description: An optional field used by Slack for endpoint verification
                  type: string
                team_id:
                  description: The team id
                  type: string
                type:
                  description: The event type
                event_id:
                  description: The event id
                  type: string
                event_time:
                  description: The event time
                  type: integer
                  format: int64
                event:
                  $ref: "#/components/schemas/SlackEvent"
              required:
                - team_id
                - type
                - event_id
                - event_time
                - event
      responses:
        200:
          $ref: "#/components/responses/ImagineResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    SlackEvent:
      description: A slack event
      properties:
        user:
          description: The user id of the sender
          type: string
        channel:
          description: The channel id the message was sent from
          type: string
        text:
          description: The content text
          type: string
        ts:
          description: The message timestamp
          type: string
        event_ts:
          description: The event timestamp
          type: string
      required:
        - user
        - channel
        - text
        - ts
        - event_ts

    Error:
      description: A generic error
      properties:
        status:
          type: integer
          format: int64
          example: 500
        code:
          $ref: '#/components/schemas/ErrorCode'
          example: ERROR
        message:
          type: string
          example: Error
      required:
        - status
        - code
        - message

    ErrorCode:
      type: string
      enum:
        - ERROR

  responses:
    ImagineResponse:
      description: Imagine Response
      content:
        application/json:
          schema:
            properties:
              challenge:
                description: An optional field used by Slack for endpoint verification
                type: string
              accepted:
                description: Determines whether the prompt was accepted
                type: boolean
            required:
              - accepted
  
    BadRequest:
        description: Bad Request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'

    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'